move examples into dedicated project
