+++ Cargo.toml
@@ -5,7 +5,7 @@ version = "0.0.5"
 authors = ["Oscar Beaumont <oscar@otbeaumont.me>"]
 edition = "2021"
 license = "MIT"
-include = ["/src", "/examples", "/LICENCE", "/README.md"]
+include = ["/src", "/LICENCE", "/README.md"]
 repository = "https://github.com/oscartbeaumont/rspc"
 documentation = "https://docs.rs/rspc/latest/rspc"
 
@@ -36,46 +36,20 @@ mac_address = ["specta/mac_address"]
 bit-vec = ["specta/bit-vec"]
 bson = ["specta/bson"]
 
-[[example]]
-name = "axum"
-required-features = ["axum"]
-
-[[example]]
-name = "cookies"
-required-features = ["axum"]
-
-[[example]]
-name = "exotic_types"
-required-features = ["uuid", "chrono"]
-
-[[example]]
-name = "global_context"
-required-features = ["axum"]
-
 [dependencies]
 specta = { version = "0.0.3", path = "./specta", features = ["serde"] }
-httpz = { path = "../../github/oscartbeaumont/httpz", optional = true } # TODO: Use crates.io version instead of file path when it's open sourced
-serde = { version = "1.0.137", features = ["derive"] }
-serde_json = "1.0.81"
-thiserror = "1.0.31"
-futures = "0.3.21"
-tokio = { version = "1.19.2", features = ["sync", "rt", "macros"] }
-tauri = { version = "1.0.4", optional = true }
-tracing = { version = "0.1.35", optional = true }
-
-[dev-dependencies]
-async-stream = "0.3.3"
-axum = "0.5.9"
-chrono = { version = "0.4.19", features = ["serde"] }
-serde = { version = "1.0.137", features = ["derive"] }
-time = "0.3.14"
-tokio = { version = "1.19.2", features = ["rt-multi-thread", "macros", "time", "sync"], default-features = false }
-tower-cookies = "0.7.0"
-tower-http = { version = "0.3.4", default-features = false, features = ["cors"] }
-uuid = { version = "1.1.2", features = ["v4", "serde"] }
+httpz = { version = "0.0.1", optional = true }
+serde = { version = "1.0.145", features = ["derive"] }
+serde_json = "1.0.85"
+thiserror = "1.0.36"
+futures = "0.3.24"
+tokio = { version = "1.21.1", features = ["sync", "rt", "macros"] }
+tauri = { version = "1.1.1", optional = true }
+tracing = { version = "0.1.36", optional = true }
 
 [workspace]
 members = [
+    "./examples",
     "./specta",
     "./specta/macros"
 ]
