+++ examples/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "example"
+version = "0.0.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+rspc = { path = "../", features = ["axum"] }
+async-stream = "0.3.3"
+axum = "0.5.16"
+chrono = { version = "0.4.22", features = ["serde"] }
+serde = { version = "1.0.145", features = ["derive"] }
+time = "0.3.14"
+tokio = { version = "1.21.1", features = ["rt-multi-thread", "macros", "time", "sync"], default-features = false }
+tower-cookies = "0.7.0"
+tower-http = { version = "0.3.4", default-features = false, features = ["cors"] }
+uuid = { version = "1.1.2", features = ["v4", "serde"] }
+serde_json = "1.0.85"
