+++ examples/src/basic.rs
@@ -1,0 +1,25 @@
+use rspc::{Router, RouterBuilder};
+
+// We merge this router into the main router in `main.rs`.
+// This router shows how to do basic queries and mutations and how they tak
+pub fn mount() -> RouterBuilder {
+    Router::new()
+        .query("version", |t| t(|_, _: ()| env!("CARGO_PKG_VERSION")))
+        .query("echo", |t| t(|_, v: String| v))
+        .query("echoAsync", |t| t(|ctx, _: i32| async move { 42 }))
+        .query("error", |t| {
+            t(|_, _: ()| {
+                Err(rspc::Error::new(
+                    rspc::ErrorCode::InternalServerError,
+                    "Something went wrong".into(),
+                )) as Result<String, rspc::Error>
+            })
+        })
+        .query("transformMe", |t| t(|_, _: ()| "Hello, world!".to_string()))
+        .mutation("sendMsg", |t| {
+            t(|_, v: String| {
+                println!("Client said '{}'", v);
+                v
+            })
+        })
+}
