+++ examples/src/cookies.rs.dis
@@ -1,0 +1,15 @@
+.query("getCookie", |t| {
+        t(|ctx, _: ()| {
+            ctx.cookies
+                .get("myDemoCookie")
+                .map(|c| c.value().to_string())
+        })
+    })
+    .mutation("setCookie", |t| {
+        t(|ctx, new_value: String| {
+            let mut cookie = Cookie::new("myDemoCookie", new_value);
+            cookie.set_expires(Some(OffsetDateTime::now_utc().add(time::Duration::DAY)));
+            cookie.set_path("/"); // Ensure you have this or it will default to `/rspc` which will cause issues.
+            ctx.cookies.add(cookie);
+        })
+    })
