+++ examples/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod basic;
+pub mod error_handling;
+pub mod selection;
+pub mod subscriptions;
