+++ specta/Cargo.toml
@@ -24,20 +24,20 @@ bit-vec = ["dep:bit-vec"]
 bson = ["dep:bson"]
 
 [dependencies]
-convert_case = "0.5.0"
+convert_case = "0.6.0"
 specta-macros = { version = "0.0.2", path = "./macros" }
 termcolor = "1.1.3"
-serde_json = { version = "1.0.81", optional = true, features = [] }
+serde_json = { version = "1.0.85", optional = true, features = [] }
 uuid = { version = "1.1.2", optional = true, features = [] }
-chrono = { version = "0.4.19", optional = true, features = [] }
-time = { version = "0.3.12", optional = true, features = [] }
+chrono = { version = "0.4.22", optional = true, features = [] }
+time = { version = "0.3.14", optional = true, features = [] }
 bigdecimal = { version = "0.3.0", optional = true, features = [] }
 rust_decimal = { version = "1.26.1", optional = true, features = [] }
 indexmap = { version = "1.9.1", optional = true, features = [] }
 ipnetwork = { version = "0.20.0", optional = true, features = [] }
-mac_address = { version = "1.1.2", optional = true, features = [] }
+mac_address = { version = "1.1.4", optional = true, features = [] }
 bit-vec = { version = "0.6.3", optional = true, features = [] }
 bson = { version = "2.4.0", optional = true, features = [] }
 
 [dev-dependencies]
-serde = { version = "1.0.140", features = ["derive"] }
+serde = { version = "1.0.145", features = ["derive"] }
