+++ specta/macros/Cargo.toml
@@ -16,7 +16,7 @@ proc-macro = true
 
 [dependencies]
 Inflector = { version = "0.11.4", default-features = false }
-proc-macro2 = "1.0.41"
-quote = "1.0.20"
-syn = { version = "1.0.98", features = ["parsing"] }
+proc-macro2 = "1.0.44"
+quote = "1.0.21"
+syn = { version = "1.0.100", features = ["parsing"] }
 termcolor = { version = "1.1.3", optional = true }
